feat: extraer funciones de cálculo a un nuevo archivo de utilidades
